feat: Services Refactor Services split into two parts: Master Service Micro Service
