+++ Integration/IntegrationClassLib/MasterServices/ComplaintResponseService.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Integration.Services
+namespace Integration.MasterServices
 {
     public class ComplaintResponseService
     {
