+++ Integration/IntegrationClassLib/MasterServices/ComplaintService.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Integration.Services
+namespace Integration.MasterServices
 {
     public class ComplaintService
     {
